% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_krishnamoorthy_ti.R
\name{add_krishnamoorthy_ti}
\alias{add_krishnamoorthy_ti}
\title{Add Krishnamoorthy Tolerance Interval Values}
\usage{
add_krishnamoorthy_ti(data, optimal_results)
}
\arguments{
\item{data}{A data frame containing at least the following columns:
\itemize{
\item By: A grouping variable
\item TI_Lower: Original lower tolerance interval bound
\item TI_Upper: Original upper tolerance interval bound
}}

\item{optimal_results}{A list of optimal results from find_optimal_point()}
}
\value{
A data frame with additional columns:
\itemize{
\item K_TI_Lower: Krishnamoorthy lower tolerance interval bound
\item K_TI_Upper: Krishnamoorthy upper tolerance interval bound
}
}
\description{
This function adds Krishnamoorthy TI values as new columns (K_TI_Lower and K_TI_Upper)
to the data frame, preserving the original TI values.
}
\examples{
\dontrun{
data <- data.frame(
  By = c("A", "B"),
  TI_Lower = c(-3, -4),
  TI_Upper = c(3, 4)
)
optimal_results <- list(
  A = list(By = "A", g1 = -2.5, g2 = 2.5),
  B = list(By = "B", g1 = -3.5, g2 = 3.5)
)
updated_data <- add_krishnamoorthy_ti(data, optimal_results)
}

}
