% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ci.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Calculate Credibility Intervals}
\usage{
calculate_ci(data, ColToAggregate, PercentForCI, ByColumn)
}
\arguments{
\item{data}{A data frame or data.table containing the data}

\item{ColToAggregate}{Column name to calculate credibility intervals for}

\item{PercentForCI}{Credibility level (between 0 and 100)}

\item{ByColumn}{Column name to group by}
}
\value{
A data frame with the following columns:
\item{By}{The grouping variable}
\item{N}{Number of observations in each group}
\item{CI_Lower}{Lower bound of the credibility interval}
\item{CI_Upper}{Upper bound of the credibility interval}
\item{Median}{Median value for each group}
}
\description{
This function calculates credibility intervals for a specified column,
grouped by another column.
}
\examples{
\dontrun{
data <- data.frame(
  Group = rep(c("A", "B"), each = 50),
  Value = c(rnorm(50, 10, 2), rnorm(50, 15, 3))
)
calculate_ci(data, "Value", 95, "Group")
}

}
