% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_difference_krishnamoorthy.R
\name{calculate_gamma_difference}
\alias{calculate_gamma_difference}
\title{Calculate Gamma Difference for Krishnamoorthy Method}
\usage{
calculate_gamma_difference(scatter_data, gP_mean, point_x_vector, gamma_actual)
}
\arguments{
\item{scatter_data}{A data frame containing at least the following columns:
\itemize{
\item By: A grouping variable
\item lower_quantile: Lower quantile values
\item upper_quantile: Upper quantile values
}}

\item{gP_mean}{Numeric, the gP_mean value for the By group}

\item{point_x_vector}{Numeric, the x-coordinate(s) to evaluate}

\item{gamma_actual}{Numeric, the target gamma (confidence level) percentage}
}
\value{
A list containing:
\itemize{
\item gamma_differences: Differences between target and achieved gamma
\item points_satisfying: Number of points satisfying the condition
\item total_points: Total number of points evaluated
}
}
\description{
This function calculates the difference between the target gamma (confidence level)
and the actual gamma achieved for a given set of points.
}
\examples{
\dontrun{
data <- data.frame(
  By = c("A", "A", "A", "A"),
  lower_quantile = c(-2, -1, -3, -2),
  upper_quantile = c(2, 3, 1, 2)
)
gP_mean_A <- 0.5
result <- calculate_gamma_difference(
  data, gP_mean_A, c(-2, -1.5, -1), 95
)
}

}
