% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_cdf_dataframes.R
\name{create_all_cdf_dataframes}
\alias{create_all_cdf_dataframes}
\title{Create CDF Data for All By-Values}
\usage{
create_all_cdf_dataframes(
  pred_dist,
  n_points = 20,
  ci_level = 90,
  log_normal = FALSE
)
}
\arguments{
\item{pred_dist}{A dataframe containing predicted distributions (output from calculate_predicted_distributions)}

\item{n_points}{Integer, number of points for CDF evaluation (default = 20)}

\item{ci_level}{Numeric, confidence interval level as percentage (default = 90)}

\item{log_normal}{Logical, whether to use log-normal transformation (default = FALSE)}
}
\value{
A list of dataframes, each containing CDF data for a specific by-value
}
\description{
This function creates CDF data for all by-values in the predicted distributions output.
}
\examples{
\dontrun{
pred_dist <- calculate_predicted_distributions(data, "fixed1", "random1",
                                             "Group", 95)
all_cdf_data <- create_all_cdf_dataframes(pred_dist)
}

}
