% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_n_inspect_data.R
\name{inspect_loaded_data}
\alias{inspect_loaded_data}
\title{Inspect Loaded Data}
\usage{
inspect_loaded_data(data)
}
\arguments{
\item{data}{A data frame to inspect}
}
\value{
The input data frame, invisibly
}
\description{
Inspect Loaded Data
}
\keyword{internal}
