% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_x_value_input.R
\name{parse_x_value_input}
\alias{parse_x_value_input}
\title{Parse X Value Input}
\usage{
parse_x_value_input(input_text)
}
\arguments{
\item{input_text}{Character string to parse, either a single number or a range in the format "\link{min,max}"}
}
\value{
A list with components:
\itemize{
\item{type}{Either "single" for a single value or "range" for a range}
\item{value}{The numeric value (if type is "single")}
\item{xmin}{The minimum value (if type is "range")}
\item{xmax}{The maximum value (if type is "range")}
}

NULL if the input cannot be parsed
}
\description{
Parses a string input representing either a single numeric value or a range \link{min,max}.
}
\examples{
parse_x_value_input("5.2")
parse_x_value_input("[2.1, 7.8]")

}
