% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{start_app_view}
\alias{start_app_view}
\title{Launch the Shiny Application}
\usage{
start_app_view(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch the Shiny Application
}
\examples{
if(interactive()){
  start_app_view()
}
}
