% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_n_inspect_data.R
\name{try_loading_methods}
\alias{try_loading_methods}
\title{Try Different Loading Methods}
\usage{
try_loading_methods(file_path, silent)
}
\arguments{
\item{file_path}{Path to the data file}

\item{silent}{Logical, whether to suppress messages}
}
\value{
A data frame containing the loaded data
}
\description{
Try Different Loading Methods
}
\keyword{internal}
